+++ src/components/Textarea/Textarea.story.js
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
-import { withKnobs, text, boolean } from '@storybook/addon-knobs';
+import { withKnobs, text, boolean, number } from '@storybook/addon-knobs';
 import { withInfo } from '@storybook/addon-info';
 
 import Textarea from './index.js';
@@ -23,6 +23,7 @@ stories
         <Textarea
           isDisabled={ boolean('isDisabled', false) }
           defaultValue='Delete this default value and see the placeholder'
+          maxLength={ number('maxLength', 250) }
           placeholder={ text('placeHolder', 'Enter a comment') }
           onBlur={ action('Textarea: onBlur') }
           onChange={ action('Textarea: onChange') }